Revert "lib: hal: usb: set correct frame for isoc ep"
This reverts commit fb5bfd6a1089ae0ad0af0fa0a41bfab19907cc5d.

If the current frame number is odd, and the UAC device is full
speed with interval 1ms, then the target number is even. So
we need to set the Even/Odd (micro)frame (EO_FrNum) field in
the DIEPCTLn/DOEPCTLn to even frame if the current frame number
is odd.

Change-Id: Id7492cbbcb970be1cab13f5bb2a8170bba33cc82
Signed-off-by: William Wu <william.wu@rock-chips.com>