lib: hal: src: acdcdigv2: add some sample rate modes
Dues to in the default flixible sample rate, we need
to reset GLB_KE quickly and may cause some broken sound
when the capture is after playback.

If some products require a fixed sample rate, we can
specify a strict sample rate and pre-configure the DIV
to avoid broken sound problem.

Change-Id: Id512398ad4160fce5f85b6a462fb3549869dc451
Signed-off-by: Xing Zheng <zhengxing@rock-chips.com>