lib: hal: pinctrl: add v2 driver support
* WHY there is v2 driver ?

The V1 pinctrl driver set registers by calculating offset for iomux,
it requires grf register offsets designed in some rules, and it will
be sad if someone out of rules.

So we need a driver in some simple rules. That is V2 pinctrl.

The V2 pinctrl driver never calculate offset, but to reference member
in GRF_REG structure. For example: GRF->GPIO1A_IOMUX_L = 0x000F0001;

* WHO should care about the new driver ?

The V2 pinctrl driver 100% keep the APIs same as V1 driver:
  HAL_PINCTRL_SetParam(bank, mPins, param)
  HAL_PINCTRL_SetIOMUX(bank, mPins, param)

Big difference is that the V1 APIs will do IOMUX SEL inside of driver
but V2 APIs never do it. The IOMUX SEL will be done in HAL_IOMUX_ API
which is part of pinctrl V2 driver.

* WHAT rule for v2 driver ?

For V2, the name of GRF structure members must in rule, such as:
Rule1:
  GPIO<_B><_P>_IOMUX_H
  GPIO<_B><_P>_IOMUX_L
  GPIO<_B><_P>_DS_H
  GPIO<_B><_P>_DS_L
  GPIO<_B><_P>_P
  GPIO<_B><_P>_IE
  GPIO<_B><_P>_SMT
    the <_B> is 0, 1, 2, 3, ... according to SoC
    the <_P> is A, B, C, D

Rule2:
  IOFUNC_CON{0,1,2,3,4,5}

Rule3:
  PMUGRF must be merged to GRF

Signed-off-by: Jianqun Xu <jay.xu@rock-chips.com>
Change-Id: I61897ad3f876a4bf60cd2c1737731426a20d6666