lib: hal: usb: fix txfifo number assignment
The DWC2 controller use the register DIEPTXFn
(Offset: 104h + (FIFO_number – 1) * 04h) to hold the size
and memory start address for each IN endpoint TxFIFOs.
According to the DWC2 databook, this register is repeated
for instantiated IN endpoint FIFOs 1 to 15.

Actually, for RK2108, it can only support FIFOs 1 to 6,
because it supports 6 IN endpoints (EP1/3/5/7/8/9-IN).
And the HAL usb driver use fixed mapping for IN endpoints
and FIFOs number.

For example, if the usb device supports 1 * UAC + 3 * HID,
and the g_usb_ep_pool is configured in the board file:

struct ep_id g_usb_ep_pool[] =
{
    { 0x0,  USB_EP_ATTR_CONTROL,    USB_DIR_INOUT,  64,   ID_ASSIGNED   },
    { 0x1,  USB_EP_ATTR_ISOC,       USB_DIR_IN,     1024, ID_UNASSIGNED },
    { 0x2,  USB_EP_ATTR_ISOC,       USB_DIR_OUT,    512,  ID_UNASSIGNED },
    { 0x3,  USB_EP_ATTR_INT,        USB_DIR_IN,     64,   ID_UNASSIGNED },
    { 0x4,  USB_EP_ATTR_INT,        USB_DIR_OUT,    64,   ID_UNASSIGNED },
    { 0x5,  USB_EP_ATTR_INT,        USB_DIR_IN,     64,   ID_UNASSIGNED },
    { 0x6,  USB_EP_ATTR_INT,        USB_DIR_OUT,    64,   ID_UNASSIGNED },
    { 0x7,  USB_EP_ATTR_INT,        USB_DIR_IN,     64,   ID_UNASSIGNED },
    { 0x8,  USB_EP_ATTR_INT,        USB_DIR_OUT,    64,   ID_UNASSIGNED },
    { 0x9,  USB_EP_ATTR_INT,        USB_DIR_IN,     64,   ID_UNASSIGNED },
    { 0xFF, USB_EP_ATTR_TYPE_MASK,  USB_DIR_MASK,   0,    ID_ASSIGNED   },
};

Without the patch, the mapping between IN endpoints
and FIFOs number:

FIFO number = IN endpoint Addr - 1

EP1-IN --> FIFO number 1
EP3-IN --> FIFO number 3
EP5-IN --> FIFO number 5
EP7-IN --> FIFO number 7
EP9-IN --> FIFO nmuber 9

For RK2108 DWC2, it fails to set the TxFIFO for EP7-IN
and EP9-IN because its DWC2 IP can can only support the
FIFOs 1 to 6.

With this patch, the mapping between IN endpoints and
and FIFOs number:

FIFO number = IN endpoint Addr >> 1

EP1-IN --> FIFO number 1
EP3-IN --> FIFO number 2
EP5-IN --> FIFO number 3
EP7-IN --> FIFO number 4
EP9-IN --> FIFO nmuber 5

Note: The EP8 and EP9 of RK2108 are bidirectional endpoints,
we suggest that set EP8 for OUT endpoint and EP9 for IN
endpoint.

Change-Id: I3158d14f1675cebd57386a6daac9af64562ad048
Signed-off-by: William Wu <william.wu@rock-chips.com>