lib: CMSIS: Device: rk3568: keep option -fomit-frame-pointer in startup_rk3568.c
If the compiler optimization level is O0, compiler will
add frame pointer code such as push/pop fp in C function IRQ_Handler,
Vectors and Reset_Handler, that will rise errors of interrupt
handle flow. So we must keep the compiler option -fomit-frame-pointer
to make sure the compiled code of function IRQ_Handler,
Vectors and Reset_Handler is what we expect it to be.

Signed-off-by: XiaoDong Huang <derrick.huang@rock-chips.com>
Change-Id: If4f820ee4c1b529c8a3a21f9f54bd4a7afb39728