lib: hal: mbox: fix the client null pointer in irq handler
HAL Mbox depends on the client to clear interrupt status and receive the
message to the registed client, so the invoker should ensure client memory
must not be released until HAL_MBOX_UnregisterClient() API is invoked.

In addition, the client pointer may be null when the channel is
unregistered, so check the interrupt status first to avoid this case.

Signed-off-by: Frank Wang <frank.wang@rock-chips.com>
Change-Id: I6203ae886ef0857951f1e1055a07cca54efabff2