lib: hal: gic: fix GIC_EnableDistributor/GIC_DisableDistributor
The GICV3 document has the following description about GICD_CLTR:
When access is Non-secure:
EnableGrp1, bit [0]
If ARE_NS == 0, then this bit is a read-write alias of the Secure GICD_CTLR.EnableGrp1NS bit.
If ARE_NS == 1, then this bit is RES0.
EnableGrp1A, bit [1]
If ARE_NS == 1, then this bit is a read-write alias of the Secure GICD_CTLR.EnableGrp1NS bit.
If ARE_NS == 0, then this bit is RES0.

So we set/clear both bit[0] and bit[1] for convenience.

Signed-off-by: XiaoDong Huang <derrick.huang@rock-chips.com>
Change-Id: I11cc687a36d258dc56890831b8cee3d7ec00fd99