lib: hal: pdm: set the pdm default gain to 24dB
Set the default gain 24dB, this parameter can get better
performance if the voice energy is lower. In other words this
can improve PDM IP SNR.

So the applicable range of this is for sound intensity below 100dB.
If you want to record stronger sound intensity, you must set
PDM gain regiter but not soft gain-controler.

Change-Id: If26ec9578b7d278bf7fd23b8970c1683c8caafd8
Signed-off-by: Jason Zhu <jason.zhu@rock-chips.com>